+++ server/routes/index.js
@@ -1,5 +1,5 @@
-const noteRoutesEmployees = require('./note_routes_employees');
-const noteRoutesSettings = require('./note_routes_settings');
+const noteRoutesEmployees = require('./routes_employees');
+const noteRoutesSettings = require('./routes_settings');
 
 module.exports = (app, db) => {
     noteRoutesSettings(app, db);
